Snake game source code
The main idea of the snake in the game is to eat and grow till it reaches its maximum size. Every feature in this game is
modifiable like the head of the snake, body, food shape, movement controls, background, etc.
Every time the score and high score get updated. The high score automatically saves the highest score each time.